feat: offline support

+++ examples/SampleApp/src/components/ChannelInfoOverlay.tsx
@@ -24,9 +24,12 @@ import {
   Avatar,
   CircleClose,
   Delete,
+  StreamCache,
   User,
   UserMinus,
   useTheme,
+  useToastContext,
+  useTranslationContext,
   vh,
   vw,
 } from 'stream-chat-react-native';
@@ -105,6 +108,9 @@ export type ChannelInfoOverlayProps = {
 
 export const ChannelInfoOverlay = (props: ChannelInfoOverlayProps) => {
   const { overlayOpacity, visible } = props;
+
+  const toast = useToastContext();
+  const { t } = useTranslationContext();
 
   const { overlay, setOverlay } = useAppOverlayContext();
   const { setData } = useBottomSheetOverlayContext();
@@ -362,11 +368,15 @@ export const ChannelInfoOverlay = (props: ChannelInfoOverlayProps) => {
                       {otherMembers.length > 1 && (
                         <TapGestureHandler
                           onHandlerStateChange={({ nativeEvent: { state } }) => {
-                            if (state === State.END) {
-                              if (clientId) {
-                                channel.removeMembers([clientId]);
+                            if (!StreamCache.getInstance().currentNetworkState) {
+                              toast.show(t('Something went wrong'), 2000);
+                            } else {
+                              if (state === State.END) {
+                                if (clientId) {
+                                  channel.removeMembers([clientId]);
+                                }
+                                setOverlay('none');
                               }
-                              setOverlay('none');
                             }
                           }}
                         >
@@ -380,21 +390,25 @@ export const ChannelInfoOverlay = (props: ChannelInfoOverlayProps) => {
                       )}
                       <TapGestureHandler
                         onHandlerStateChange={({ nativeEvent: { state } }) => {
-                          if (state === State.END) {
-                            setData({
-                              confirmText: 'DELETE',
-                              onConfirm: () => {
-                                channel.delete();
-                                setOverlay('none');
-                              },
-                              subtext: `Are you sure you want to delete this ${
-                                otherMembers.length === 1 ? 'conversation' : 'group'
-                              }?`,
-                              title: `Delete ${
-                                otherMembers.length === 1 ? 'Conversation' : 'Group'
-                              }`,
-                            });
-                            setOverlay('confirmation');
+                          if (!StreamCache.getInstance().currentNetworkState) {
+                            toast.show(t('Something went wrong'), 2000);
+                          } else {
+                            if (state === State.END) {
+                              setData({
+                                confirmText: 'DELETE',
+                                onConfirm: () => {
+                                  channel.delete();
+                                  setOverlay('none');
+                                },
+                                subtext: `Are you sure you want to delete this ${
+                                  otherMembers.length === 1 ? 'conversation' : 'group'
+                                }?`,
+                                title: `Delete ${
+                                  otherMembers.length === 1 ? 'Conversation' : 'Group'
+                                }`,
+                              });
+                              setOverlay('confirmation');
+                            }
                           }
                         }}
                       >
