+++ package/src/components/Message/Message.tsx
@@ -59,6 +59,7 @@ import {
   useOverlayContext,
 } from '../../contexts/overlayContext/OverlayContext';
 import { useTheme } from '../../contexts/themeContext/ThemeContext';
+import { useToastContext } from '../../contexts/toastContext/ToastContext';
 import { ThreadContextValue, useThreadContext } from '../../contexts/threadContext/ThreadContext';
 import {
   TranslationContextValue,
@@ -77,6 +78,7 @@ import {
   UserDelete,
 } from '../../icons';
 import { triggerHaptic } from '../../native';
+import { StreamCache } from '../../StreamCache';
 import { emojiRegex } from '../../utils/utils';
 
 import type { Attachment, MessageResponse, Reaction } from 'stream-chat';
@@ -359,6 +361,8 @@ const MessageWithContext = <
     threadReply: threadReplyProp,
     updateMessage,
   } = props;
+
+  const toast = useToastContext();
 
   const {
     theme: {
@@ -673,13 +677,17 @@ const MessageWithContext = <
       ? null
       : {
           action: () => async () => {
-            setOverlay('none');
-            if (message.user?.id) {
-              if (handleBlock) {
-                handleBlock(message);
-              }
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              setOverlay('none');
+              if (message.user?.id) {
+                if (handleBlock) {
+                  handleBlock(message);
+                }
 
-              await handleToggleBanUser();
+                await handleToggleBanUser();
+              }
             }
           },
           icon: <UserDelete pathFill={grey} />,
@@ -693,11 +701,15 @@ const MessageWithContext = <
       : {
           // using depreciated Clipboard from react-native until expo supports the community version or their own
           action: () => {
-            setOverlay('none');
-            if (handleCopy) {
-              handleCopy(message);
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              setOverlay('none');
+              if (handleCopy) {
+                handleCopy(message);
+              }
+              Clipboard.setString(message.text || '');
             }
-            Clipboard.setString(message.text || '');
           },
           icon: <Copy pathFill={grey} />,
           title: t('Copy Message'),
@@ -709,28 +721,32 @@ const MessageWithContext = <
       ? null
       : {
           action: () => {
-            setOverlay('alert');
-            if (message.id) {
-              Alert.alert(
-                t('Delete Message'),
-                t('Are you sure you want to permanently delete this message?'),
-                [
-                  { onPress: () => setOverlay('none'), text: t('Cancel') },
-                  {
-                    onPress: async () => {
-                      setOverlay('none');
-                      if (handleDelete) {
-                        handleDelete(message);
-                      }
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              setOverlay('alert');
+              if (message.id) {
+                Alert.alert(
+                  t('Delete Message'),
+                  t('Are you sure you want to permanently delete this message?'),
+                  [
+                    { onPress: () => setOverlay('none'), text: t('Cancel') },
+                    {
+                      onPress: async () => {
+                        setOverlay('none');
+                        if (handleDelete) {
+                          handleDelete(message);
+                        }
 
-                      await handleDeleteMessage();
+                        await handleDeleteMessage();
+                      },
+                      style: 'destructive',
+                      text: t('Delete'),
                     },
-                    style: 'destructive',
-                    text: t('Delete'),
-                  },
-                ],
-                { cancelable: false },
-              );
+                  ],
+                  { cancelable: false },
+                );
+              }
             }
           },
           icon: <Delete pathFill={accent_red} />,
@@ -744,11 +760,15 @@ const MessageWithContext = <
       ? null
       : {
           action: () => {
-            setOverlay('none');
-            if (handleEdit) {
-              handleEdit(message);
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              setOverlay('none');
+              if (handleEdit) {
+                handleEdit(message);
+              }
+              handleEditMessage();
             }
-            handleEditMessage();
           },
           icon: <Edit pathFill={grey} />,
           title: t('Edit Message'),
@@ -760,50 +780,54 @@ const MessageWithContext = <
       ? null
       : {
           action: () => {
-            setOverlay('alert');
-            if (message.id) {
-              Alert.alert(
-                t('Flag Message'),
-                t(
-                  'Do you want to send a copy of this message to a moderator for further investigation?',
-                ),
-                [
-                  { onPress: () => setOverlay('none'), text: t('Cancel') },
-                  {
-                    onPress: async () => {
-                      try {
-                        if (handleFlag) {
-                          handleFlag(message);
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              setOverlay('alert');
+              if (message.id) {
+                Alert.alert(
+                  t('Flag Message'),
+                  t(
+                    'Do you want to send a copy of this message to a moderator for further investigation?',
+                  ),
+                  [
+                    { onPress: () => setOverlay('none'), text: t('Cancel') },
+                    {
+                      onPress: async () => {
+                        try {
+                          if (handleFlag) {
+                            handleFlag(message);
+                          }
+                          await client.flagMessage(message.id);
+                          Alert.alert(
+                            t('Message flagged'),
+                            t('The message has been reported to a moderator.'),
+                            [
+                              {
+                                onPress: () => setOverlay('none'),
+                                text: t('Dismiss'),
+                              },
+                            ],
+                          );
+                        } catch (err) {
+                          Alert.alert(
+                            t('Something went wrong'),
+                            t("The operation couldn't be completed."),
+                            [
+                              {
+                                onPress: () => setOverlay('none'),
+                                text: t('Dismiss'),
+                              },
+                            ],
+                          );
                         }
-                        await client.flagMessage(message.id);
-                        Alert.alert(
-                          t('Message flagged'),
-                          t('The message has been reported to a moderator.'),
-                          [
-                            {
-                              onPress: () => setOverlay('none'),
-                              text: t('Dismiss'),
-                            },
-                          ],
-                        );
-                      } catch (err) {
-                        Alert.alert(
-                          t('Something went wrong'),
-                          t("The operation couldn't be completed."),
-                          [
-                            {
-                              onPress: () => setOverlay('none'),
-                              text: t('Dismiss'),
-                            },
-                          ],
-                        );
-                      }
+                      },
+                      text: t('Flag'),
                     },
-                    text: t('Flag'),
-                  },
-                ],
-                { cancelable: false },
-              );
+                  ],
+                  { cancelable: false },
+                );
+              }
             }
           },
           icon: <MessageFlag pathFill={grey} />,
@@ -814,11 +838,15 @@ const MessageWithContext = <
       ? selectReaction
         ? selectReaction(message)
         : async (reactionType: string) => {
-            if (handleReactionProp) {
-              handleReactionProp(message, reactionType);
-            }
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              if (handleReactionProp) {
+                handleReactionProp(message, reactionType);
+              }
 
-            await handleToggleReaction(reactionType);
+              await handleToggleReaction(reactionType);
+            }
           }
       : undefined;
 
@@ -828,13 +856,17 @@ const MessageWithContext = <
       ? null
       : {
           action: async () => {
-            setOverlay('none');
-            if (message.user?.id) {
-              if (handleMute) {
-                handleMute(message);
-              }
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              setOverlay('none');
+              if (message.user?.id) {
+                if (handleMute) {
+                  handleMute(message);
+                }
 
-              await handleToggleMuteUser();
+                await handleToggleMuteUser();
+              }
             }
           },
           icon: <Mute pathFill={grey} />,
@@ -847,11 +879,15 @@ const MessageWithContext = <
       ? null
       : {
           action: () => {
-            setOverlay('none');
-            if (handleQuotedReply) {
-              handleQuotedReply(message);
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              setOverlay('none');
+              if (handleQuotedReply) {
+                handleQuotedReply(message);
+              }
+              handleQuotedReplyMessage();
             }
-            handleQuotedReplyMessage();
           },
           icon: <CurveLineLeftUp pathFill={grey} />,
           title: t('Reply'),
@@ -863,13 +899,17 @@ const MessageWithContext = <
       ? null
       : {
           action: async () => {
-            setOverlay('none');
-            const messageWithoutReservedFields = removeReservedFields(message);
-            if (handleRetry) {
-              handleRetry(messageWithoutReservedFields);
-            }
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              setOverlay('none');
+              const messageWithoutReservedFields = removeReservedFields(message);
+              if (handleRetry) {
+                handleRetry(messageWithoutReservedFields);
+              }
 
-            await handleResendMessage();
+              await handleResendMessage();
+            }
           },
           icon: <SendUp pathFill={accent_blue} />,
           title: t('Resend'),
@@ -881,11 +921,15 @@ const MessageWithContext = <
       ? null
       : {
           action: () => {
-            setOverlay('none');
-            if (handleThreadReply) {
-              handleThreadReply(message);
+            if (!StreamCache.getInstance().currentNetworkState) {
+              toast.show(t('Something went wrong'), 2000);
+            } else {
+              setOverlay('none');
+              if (handleThreadReply) {
+                handleThreadReply(message);
+              }
+              onOpenThread();
             }
-            onOpenThread();
           },
           icon: <ThreadReply pathFill={grey} />,
           title: t('Thread Reply'),
